airflow/requirements.txt
requests
psycopg2-binary
pandas
beautifulsoup4==4.9.3

openpyxl